requirements.txt
markovify==0.8.0
names==0.3.0
Pillow==7.1.2
SQLAlchemy==1.3.16
